backend/requirements/base.txt
Django==3.1.7
djangorestframework==3.12.4
psycopg2-binary==2.8.6